// backend/src/modules/task/dto/res.edittask.dto.ts
export class ResEditTaskDto {
  message!: string;
  task!: {
    id: number;
    title: string;
    description?: string | null;
    dueDate?: Date | null;
    status: string;
    taskOrder: number;
    createdBy: number;
    assignedTo?: number | null;
    createdAt: Date;
    updatedAt: Date;
  };
}